+++ srcs/str/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/24 17:43:23 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/11/05 11:19:03 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/11/18 00:43:17 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ size_t	ft_strlcat(char *restrict dest,
 	n = dstsize;
 	while (n-- && *dst++)
 		;
-	dlen = dst - dest;
+	dlen = (size_t)(dst - dest);
 	if (!(n = dstsize - dlen))
 		return (dlen + ft_strlen(sc));
 	while (*sc != '\0')
@@ -38,5 +38,5 @@ size_t	ft_strlcat(char *restrict dest,
 		++sc;
 	}
 	*dst = '\0';
-	return (dlen + (sc - src));
+	return (dlen + (size_t)(sc - src));
 }
